+++ tests/test_iter_patterns.py
@@ -7,7 +7,7 @@ to see if any regressions might happen in the future).
 
 Author, Copyright, and License
 ------------------------------
-Copyright (c) 2022-2023 Hauke Daempfling (haukex@zero-g.net)
+Copyright (c) 2022-2024 Hauke Daempfling (haukex@zero-g.net)
 at the Leibniz Institute of Freshwater Ecology and Inland Fisheries (IGB),
 Berlin, Germany, https://www.igb-berlin.de/