+++ docs/changelog.rst
@@ -4,7 +4,7 @@ Changelog for ``igbpyutils``
 Changelog
 ---------
 
-v0.8.0 - *not yet released*
+v0.8.0 - Fri Oct 11 2024
     - **REMOVED** ``gen-coveragerc`` and **deprecated** :func:`~igbpyutils.dev.generate_coveragerc`
         - Use https://pypi.org/project/coverage-simple-excludes/ instead.
     - **Renamed** ``igbpyutils.dev.check_script_vs_lib`` to :func:`igbpyutils.dev.script_vs_lib.check_script_vs_lib`